finished permissions page most of the logic associated with adding and deleting items or permissions for users

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
 const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
     async createItem( parent, args, ctx, info ){
@@ -43,9 +44,16 @@ const Mutations = {
     async deleteItem( parent, args, ctx, info ) {
         const where = { id: args.id };
         // 1. Find the item
-        const item = await ctx.db.query.item({ where }, `{ id, title }`);
+        const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
         // 2. Check if they own that item, or have the permissions
-        // -TODO-
+        const ownsItem = item.user.id === ctx.request.useriD;
+        const hasPermissions = ctx.request.user.permissions.some(
+            permission => ['ADMIN', 'ITEMDELETE'].includes(permission)
+        );
+
+        if ( !ownsItem && !hasPermissions ) {
+            throw new Error('You do not have permission to do this!');
+        }
         // 3. Delete it!
         return ctx.db.mutation.deleteItem({ where }, info);
     },
@@ -171,6 +179,31 @@ const Mutations = {
         return updatedUser;
     },
 
+    async updatePermissions( parent, args, ctx, info ) {
+        // 1. Check if they are logged in
+        if(!ctx.request.userId) {
+            throw new Error('You must be logged in!');
+        }
+        // 2. Query the current user
+        const currentUser = await ctx.db.query.user({
+            where: {
+                id: ctx.request.userId
+            }
+        }, info );
+        // 3. Check if they have permissions to do this
+        hasPermission( currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+        // 4. Update the permissions
+        return ctx.db.mutation.updateUser({
+            data: {
+                permissions: {
+                    set: args.permissions
+                },
+            },
+            where: {
+                id: args.userId
+            },
+        }, info );
+    },
 };
 
 module.exports = Mutations;
